refactor structure to put related classes in their own modules and packages
