+++ gui/refactor_game/data/log.py
@@ -1,0 +1,5 @@
+class LogItem():
+    def __init__(self, move, result_board, time_taken):
+        self.move = move
+        self.board = result_board
+        self.time_taken = time_taken
